pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codesense-ai"
version = "0.1.0"
description = "AI-Powered Code Review Assistant with GitHub Integration"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Amr Hassan", email = "amrgabeerr20@gmail.com" }
]
keywords = ["ai", "code-review", "llm", "github", "static-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"

dependencies = [
    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",
    
    # API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # LLM
    "langchain>=0.1.0",
    "langchain-groq>=0.0.1",
    "langchain-openai>=0.0.5",
    
    # GitHub
    "PyGithub>=2.1.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]

[project.scripts]
codesense = "codesense.cli.app:app"

[project.urls]
Homepage = "https://github.com/amrgaberM/codesense-ai"
Documentation = "https://github.com/amrgaberM/codesense-ai#readme"
Repository = "https://github.com/amrgaberM/codesense-ai"
Issues = "https://github.com/amrgaberM/codesense-ai/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/codesense"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]  # line too long (handled by black)

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=codesense --cov-report=term-missing"

[tool.coverage.run]
source = ["src/codesense"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
]
